Homework ver.2

+++ homework.py
@@ -26,7 +26,7 @@ class InfoMessage:
 class Training:
     """Базовый класс тренировки."""
     LEN_STEP: float = 0.65
-    M_IN_KM: float = 1000
+    M_IN_KM: int = 1000
     MIN_IN_H: int = 60
 
     def __init__(self,
